pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canada-retirement-planner"
version = "0.1.0"
description = "AI-assisted Canadian retirement planning MVP for Ontario 2026 assumptions."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "openai>=1.99.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.2",
  "typer>=0.12.5",
  "uvicorn[standard]>=0.30.6",
  "weasyprint>=62.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.4",
]

[project.scripts]
retirement = "retirement_calculator.cli:app"

[tool.setuptools.packages.find]
where = ["backend"]

[tool.setuptools.package-data]
retirement_calculator = ["templates/*.html"]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
